pyproject.toml
[tool.poetry]
name = "chatgpt_robusta_actions"
version = "0.0.44"
description = ""
authors = []

[tool.poetry.dependencies]
openai = "^0.27.2"
cachetools = "^5.3.0"
python = ">=3.7.1,<4.0"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
robusta-cli = "^0.10.14"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
